fix(abi): lastKeyCode is uint8_t, not i16 -- keyboard state machine corruption
lastKeyCode is defined u8 (frontier_c47.zig:305) and declared uint8_t upstream
(c47.h:428), but keyboard_state_runtime bound it as i16, and the keyboard state
machine writes through that binding: keyboard_state_shared / keyboard_state do
`runtime.lastKeyCode = @intCast(key + 20)` etc. As an i16 store that writes two
bytes over the one-byte symbol, clobbering the adjacent global's low byte on
every keypress (little-endian). It was invisible because the keyboard_state
parity fake ALSO declared int16_t -- the whole parity world agreed on the wrong
width while real firmware uses uint8_t.

Reconcile lastKeyCode to uint8_t across the owner binding and the parity harness
(fake runtime, c47.h, parity_runtime). All written key codes (20..84) and the
test value 43 fit u8; the parity snapshot field stays int16_t (harmless promotion
of the now-uint8_t global).

Verified: zig build sim and zig build test exit 0 (9674 testSuite),
keyboard_state_parity green, fmt and Phase-I c-dependency clean.

+++ zig_build/tests/keyboard_state/c47.h
@@ -91,7 +91,7 @@ enum {
 
 extern uint8_t calcMode;
 extern int16_t itemToBeAssigned;
-extern int16_t lastKeyCode;
+extern uint8_t lastKeyCode;
 extern tam_state_t tam;
 extern calcKey_t kbd_std[37];
 extern calcKey_t kbd_usr[37];
