+++ zig_build/tests/keyboard_state/keyboard_state_parity_runtime.zig
@@ -65,7 +65,7 @@ pub const SCRUPD_SKIP_STACK_ONE_TIME: u8 = 0x20;
 
 pub extern var calcMode: u8;
 pub extern var itemToBeAssigned: i16;
-pub extern var lastKeyCode: i16;
+pub extern var lastKeyCode: u8;
 pub extern var tam: tam_state_t;
 pub extern var kbd_usr: [37]calcKey_t;
 pub extern var currentFlgScr: u8;
